Adds discover music video section to homepage
Implements a new section on the homepage to display music videos.

Fetches the latest music videos from the database and displays them in a carousel.
Includes a main video player and a navigation list of thumbnails.
Supports playing and stopping videos, displaying video information and YouTube thumbnails.

Introduces a new composable method to fetch the latest MVs.